coordinator/requirements.txt
flask==2.0.2
flasgger==0.9.5
SQLAlchemy==1.4.26
pika==1.2.0
Flask-OAuthlib==0.9.6
psycopg2-binary==2.9.1
flask-login==0.5.0
marshmallow_sqlalchemy==0.26.1
flask-cors==3.0.10
Werkzeug==2.0.0
jinja2==3.0.3
python-powerdns==2.1.0